requirements.txt
numpy==1.24.1
pytorch-metric-learning==2.1.0
torch-geometric==2.3.0
torchmetrics==0.11.4
tqdm==4.64.1